refactor: migra sistema de carrinho para PedidoProduto e BigDecimal
- Altera valorTotal de int para BigDecimal na classe Pedido
- Atualiza validação de valorTotal para usar compareTo() com BigDecimal
- Substitui CarrinhoItem por PedidoProduto no sistema de carrinho
- Adiciona obterPedidoTemporario() para gerenciar pedido na sessão
- Implementa atualizarQuantidade() para modificar itens do carrinho
- Valida login ao acessar carrinho (redireciona para index com modal)
- Adiciona validação de carrinho vazio no checkout
- Usa BigDecimal em todos os cálculos monetários para maior precisão
- Adiciona totalCarrinhoFormatado nos models para exibição formatada
- Remove arquivos obsoletos

+++ EcommerceBD/src/main/java/com/mycompany/ecommercebd/model/DAO/NotaFiscalDAO.java
@@ -188,7 +188,7 @@ public class NotaFiscalDAO extends EntidadeBaseDAO<NotaFiscal> {
                     p.setId(rs.getLong("ID_pedido"));
                     p.setData(rs.getTimestamp("Data_do_pedido").toLocalDateTime());
                     p.setStatus(rs.getString("Status"));
-                    p.setValorTotal(rs.getInt("Valor_total"));
+                    p.setValorTotal(rs.getBigDecimal("Valor_total"));
 
                     Long Idcliente = rs.getLong("fk_Cliente_id");
                     ClienteDAO cd = new ClienteDAO(connection);
